moved logging outside Manticore
